refactor(cli): migrate construct command into Command pipeline and split itemUse handler
