+++ packs/BP/scripts/consts.js
@@ -1,0 +1,2 @@
+export const PACK_IDENTIFIER = 'construct';
+export const MENU_ITEM = 'construct:menu';
